Add a dummy animation to set ispony
This lets servers set the ispony variable when sending a skin (IDK what the property is called in the protocol, but NukkitX calls it "SkinResourcePatch", which is undocumented but appears to let you patch certain parts of the player entity. In my testing, variables can't be directly set, but animations can be switched out.)
It's a bit janky but I haven't discovered a better way to do this yet.